Mowing Time
This is a fairly base program that does some simple calculations to show
the time it would take to mow a lawn. The house area and property area
are given as values, and I added the option to change the rate depending
if a push mower or a driven mower is being used.